move product-content index.js file to product folder for better folder structure
